requirements.txt
httpx[socks]>=0.27
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
APScheduler>=3.10
jinja2>=3.1
python-dotenv>=1.0
rich>=13.7
scipy>=1.11
tenacity>=8.2
beautifulsoup4>=4.12
lxml>=5.0
streamlit>=1.32
plotly>=5.20
altair>=5.2
